Implement a pseudo-random number generator.
This is a really primitive way of generating randomness, it just
collects the number of ticks spent waiting at the title screen for the
user to press space and uses that number to provide an initial seed to
the RNG. From there the RNG just adds whatever value it randomly
selects, each time it is used, to the entropy value.

It's not going to win any crypto awards, but it should be good enough to
shuffle the "bag" of tetris pieces.