Implement basic rotation.
The rotation pretty much works for the most basic case, where the
rotation doesn't hit a wall, and "kicks" are not handled yet.

In order to handle wall collisions, I'll need to come up with a nice way
to determine wall clearance. This would be a lot easier if I was using
an x, y coordinate scheme but switching away from the current linear
index scheme would come with a lot of other downsides, so I'd rather
not.