Add detection of completed lines.
We can now detect when lines have been completed after a piece locks
down, and display the total number of completed lines on the screen.
Clearing out completed lines is still TODO.

Also add a 3-second start game countdown, and plug some memory leaks. It
turns out just feeding a string literal directly to Output.printString()
is a bad idea because it will never get disposed. We need to explicitly
dispose of all String literals, so I've added a Util function for that.